Fix workflow bundler warn due to non-workflow safe import in user-metadata.ts
